Spring API with ExceptionHandle

+++ Spring-API-Demo/src/main/java/com/example/service/ProductServiceImpl.java
@@ -1,13 +1,15 @@
 package com.example.service;
 
 import com.example.dto.ProductDto;
+import com.example.exception.ProductAlreadyExistsException;
+import com.example.exception.ProductNotFoundException;
 import com.example.mapper.ProductMapper;
 import com.example.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+
 
 @Service
 public class ProductServiceImpl implements ProductService{
@@ -18,21 +20,48 @@ public class ProductServiceImpl implements ProductService{
 
     @Override
     public List<ProductDto> getAll() {
+        if(mapper.modelsToDtos(repo.findAll()).isEmpty()){
+            throw new ProductNotFoundException();
+        }
         return mapper.modelsToDtos(repo.findAll());
     }
 
     @Override
     public void delete(Integer id) {
+        if (repo.findById(id).isEmpty()){
+            throw new ProductNotFoundException();
+        }
         repo.deleteById(id);
     }
 
     @Override
     public void save(ProductDto productDto) {
+        if (repo.existsById(productDto.getId())){
+            throw new ProductAlreadyExistsException();
+        }
         repo.save(mapper.DtoToModel(productDto));
     }
 
     @Override
-    public Optional<ProductDto> findById(Integer id) {
-        return Optional.of(mapper.modelToDto(repo.findById(id).get()));
+    public ProductDto getProductById(Integer id){
+          if(repo.findById(id).isEmpty()){
+              throw new ProductNotFoundException();
+          } else {
+              return mapper.modelToDto(repo.findById(id).get());
+          }
+    }
+    @Override
+    public void update(Integer id, ProductDto productDto){
+        if (repo.findById(id).isEmpty()) {
+            throw new ProductNotFoundException();
+        }else if(repo.existsById(productDto.getId()) && productDto.getId() != (repo.findById(id).get().getId())){
+            throw new ProductAlreadyExistsException();
+        }
+        ProductDto list = mapper.modelToDto(repo.findById(id).get());
+        list.setName(productDto.getName());
+        list.setPrice(productDto.getPrice());
+        repo.save(mapper.DtoToModel(list));
+//        repo.deleteById(id);
+//        repo.save(mapper.DtoToModel(productDto));
     }
 }
